Bump libtheora to 1.2.0 with Windows configure diagnostics

+++ recipe/build.sh
@@ -1,11 +1,15 @@
 #!/bin/bash
 
+set -x
+
 export C_INCLUDE_PATH=${PREFIX}/include
-export LDFLAGS="-L${PREFIX}/lib"
+export LDFLAGS="$LDFLAGS -L${PREFIX}/lib"
 export PKG_CONFIG_PATH=${PREFIX}/lib/pkgconfig
 
-rm ./config.sub
-./autogen.sh
+if [[ `uname` != MINGW* ]]; then
+  rm ./config.sub
+  ./autogen.sh
+fi
 
 # -fforce-addr is not supported in clang
 if [ `uname` == Darwin ]; then
@@ -14,7 +18,19 @@ if [ `uname` == Darwin ]; then
     export LDFLAGS="${LDFLAGS} -Wl,-rpath,$PREFIX/lib"
 fi
 
-./configure --prefix=${PREFIX} --disable-examples --disable-spec
+export CONFIGURE_ADDITIONAL_OPTIONS=""
+
+if [[ `uname` == MINGW* ]]; then
+  # GCC style assembly is not supported by clang on Windows, so let's disable it for now
+  # if you really need the fastest possible libtheora on Windows, feel free to open an issue
+  export CONFIGURE_ADDITIONAL_OPTIONS="$CONFIGURE_ADDITIONAL_OPTIONS --disable-asm"
+fi
+
+./configure --prefix=${PREFIX} --disable-examples --disable-spec ${CONFIGURE_ADDITIONAL_OPTIONS}
+
+# As documented in https://github.com/conda-forge/autotools_clang_conda-feedstock/blob/cb241060f5d8adcd105f3b2e8454a8ad4d70f08f/recipe/meta.yaml#L58C1-L58C60
+[[ `uname` == MINGW* ]] && patch_libtool
+
 make
 
 if [[ "$CONDA_BUILD_CROSS_COMPILATION" != "1" ]]; then
@@ -22,3 +38,9 @@ if [[ "$CONDA_BUILD_CROSS_COMPILATION" != "1" ]]; then
 fi
 
 make install
+
+# Delete static libraries (per CFEP-18)
+rm -rf $PREFIX/lib/libtheor*.a
+
+# Delete docs
+rm -rf $PREFIX/share/doc/libtheor*
